langchain_examples/pyproject.toml
[project]
name = "langchain_examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "langchain==0.2.11",
    "langchain-core==0.2.43",
    "langchain-anthropic==0.1.20",
    "langchain-community==0.2.10",
    "langchain-experimental==0.0.62",
    "langchain-huggingface==0.0.3",
    "langchain-chroma>=0.1.4",
    "pypdf>=4.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.7.0",
    "pypdf2>=3.0.1",
    "lark==1.3.1",
]
